integrate josh's new work.  It works!  #19 is just about solved!
