+++ Makefile
@@ -3,13 +3,15 @@ all : lib data_recorder test calibrate
 CFLAGS:=-Iinclude -fPIC -g -Os -Iredist -flto
 LDFLAGS:=-lpthread -lusb-1.0 -lz -lX11 -flto
 
+# unused: redist/crc32.c
+
 test : test.c lib/libsurvive.so redist/os_generic.o
 	gcc -o $@ $^ $(LDFLAGS) $(CFLAGS)
 
 data_recorder : data_recorder.c lib/libsurvive.so redist/os_generic.o redist/DrawFunctions.o redist/XDriver.o
 	gcc -o $@ $^ $(LDFLAGS) $(CFLAGS)
 
-calibrate :  calibrate.c lib/libsurvive.so redist/os_generic.c redist/DrawFunctions.c redist/XDriver.c redist/crc32.c
+calibrate :  calibrate.c lib/libsurvive.so redist/os_generic.c redist/DrawFunctions.c redist/XDriver.c
 	gcc -o $@ $^ $(LDFLAGS) $(CFLAGS)
 
 lib:
